project_root/database.py
import os
from dotenv import load_dotenv
from flask_sqlalchemy import SQLAlchemy
from sqlalchemy.orm import DeclarativeBase
import pymysql

# Load environment variables
load_dotenv()

# Use PyMySQL as the MySQL driver
pymysql.install_as_MySQLdb()

# Create the base class for SQLAlchemy models
class Base(DeclarativeBase):
    pass

# Initialize SQLAlchemy
db = SQLAlchemy(model_class=Base)